requirements.txt
pytest>=7.0.0
pytest-mock>=3.10.0
kafka-python>=2.0.2
pandas>=1.3.0
pyarrow>=5.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0